Move settings into its own module
In preparation to expand settings to include daemon specific things (e.g.
ports)
